Remove validation.py test file from production src, fix {tmp_data} folder creation in unit tests
